requirements.txt
# Core ML stack
tensorflow==2.15.0
keras==2.15.0
numpy==1.24.4
pandas==2.2.2
scikit-learn==1.4.2
scipy==1.11.4
albumentations==1.4.7
opencv-python==4.9.0.80
Pillow==10.3.0
imagehash==4.3.1
umap-learn==0.5.6
matplotlib==3.8.4
seaborn==0.13.2
plotly==5.20.0
tqdm==4.66.2
rich==13.7.1
joblib==1.3.2
# Desktop UI
kivy==2.3.0
kivymd==1.1.1
# Utilities
tensorflow-io-gcs-filesystem==0.34.0
pyyaml==6.0.1
pytest==8.1.1
